Use serde_derive.
Now that Rust 1.15 is out. 🎉
