+++ src/lib.rs
@@ -8,6 +8,8 @@ extern crate lettre;
 extern crate mustache;
 extern crate redis;
 extern crate serde;
+#[macro_use]
+extern crate serde_derive;
 extern crate serde_json;
 extern crate time;
 extern crate url;
